final uas

+++ routes/web.php
@@ -5,6 +5,7 @@ use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\AdminController;
 use App\Http\Controllers\MentorController;
 use App\Http\Controllers\StaffController;
+use App\Http\Controllers\StaffReController;
 use App\Http\Controllers\InstansiController;
 use App\Http\Controllers\PembimbingController;
 use App\Http\Controllers\PengajuanController;
@@ -34,7 +35,7 @@ Route::get('/', function () {
 
 
 Route::get('send-mail', [MailController::class, 'index']);
-
+Route::post('/pendaftaran-instansi',[InstansiReController::class,'storeDaftar'])->name('daftarInstansi');
 
 
 Route::middleware('auth')->group(function () {
@@ -50,41 +51,35 @@ Route::middleware(['auth', 'role:admin'])->group(function () {
 
     //Dashboard Admin
     Route::get('admin/dashboard', [AdminController::class, 'AdminDashboard'])->name('admin.dashboard');
-
     Route::get('admin/logout', [AdminController::class, 'AdminLogout'])->name('admin.logout');
 
     Route::get('admin/profile', [AdminController::class, 'AdminProfile'])->name('admin.profile');
     Route::post('admin/profile/store', [AdminController::class, 'AdminProfileStore'])->name('admin.profile.store');
-
-
     Route::get('admin/change/password', [AdminController::class, 'AdminChangePassword'])->name('admin.change.password');
     Route::post('admin/update/password', [AdminController::class, 'AdminUpdatePassword'])->name('admin.update.password');
-    Route::get('admin/create/data', [AdminController::class, 'AdminCreateData'])->name('admin.create.data');
-    Route::post('admin/data/store', [AdminController::class, 'AdminDataStore'])->name('admin.data.store');
-    Route::get('admin/edit/data', [AdminController::class, 'AdminEditData'])->name('admin.edit.data');
-    Route::delete('admin/data/delete', [AdminController::class, 'AdminDeleteUser'])->name('admin.delete.user');
+
     Route::post('admin/data/{type}/{id}',[AdminController::class, 'updateStatusAktif'])->name('admin.editStatus');
 
-
-    // Route::get('admin/datamaster', [AdminController::class, 'AdminDataMaster'])->name('admin.DataMaster');
-    // Route::get('admin/instansi', [AdminController::class, 'AdminInstansi'])->name('admin.instansi');
-    Route::get('admin/instansi/{id}', [AdminController::class, 'editInstansi'])->name('admin.editInstansi');
-    // Route::post('admin/instansi/{id}',[AdminController::class, 'updateInstansi'])->name('admin.updateInstansi');
-    // Route::post('admin/instansi', [AdminController::class, 'insertInstansi'])->name('insertInstansi');
-    Route::delete('admin/instansi/{id}', [AdminController::class,'destroyInstansi'])->name('deleteInstansi');
-
-    Route::delete('user/delete/{id}',[UserReController::class,'destroyUser'])->name('deleteUser');
-
-
     Route::get('admin/tim/get-pembimbing/{id}',[TimController::class,'getPembimbingSiswa']);
-    // Route::post('admin/tim/get-siswa',[TimController::class,'getSiswa'])->name('tim.getSiswa');
+    Route::get('admin/pengajuan',[AdminController::class,'indexPengajuan'])->name('indexPengajuan');
+    Route::get('admin/pengajuan/{id}',[AdminController::class,'detailPengajuan'])->name('detailPengajuan');
+    Route::post('admin/pengajuan/{id}',[AdminController::class,'updatePengajuan'])->name('updatePengajuan');
+
+    Route::delete('admin/instansi/delete{id}', [AdminController::class,'destroyInstansi'])->name('deleteInstansi');
+    Route::delete('admin/user/delete/{id}',[UserReController::class,'destroyUser'])->name('deleteUser');
+    Route::delete('admin/instansi/delete/{id}',[PengajuanController::class,'destroyInstansi'])->name('deleteInstansi');
+    Route::delete('admin/pembimbing/delete/{id}',[PembimbingController::class,'destroyPembimbing'])->name('deletePembimbing');
+    Route::delete('admin/siswa/delete/{id}',[SiswaController::class,'destroySiswa'])->name('deleteSiswa');
+    Route::delete('admin/tim/delete/{id}',[TimController::class,'destroyTim'])->name('deleteTim');
+    Route::delete('admin/staff/delete/{id}',[StaffReController::class,'destroyStaff'])->name('deleteStaff');
+    
 
     Route::resource('admin/user', UserReController::class);
     Route::resource('admin/instansi', InstansiReController::class);
     Route::resource('admin/pembimbing', PembimbingController::class);
     Route::resource('admin/siswa', SiswaController::class);
     Route::resource('admin/tim', TimController::class);
-    Route::resource('admin/staff', StaffController::class);
+    Route::resource('admin/staff', StaffReController::class);
 
 });
 
@@ -96,9 +91,11 @@ Route::middleware(['auth', 'role:mentor'])->group(function () {
     Route::get('mentor/create/absensi', [MentorController::class, 'MentorCreateAbsensi'])->name('mentor.create.absensi');
 
     Route::get('mentor/jurnal', [MentorController::class, 'MentorJurnal'])->name('mentor.jurnal');
-
+    Route::post('mentor/jurnal', [MentorController::class, 'MentorJurnalPost'])->name('mentor.jurnalpost');
+    Route::get('mentor/jurnals', [MentorController::class, 'CreateMateri'])->name('create.materi');
     Route::get('mentor/logout', [MentorController::class, 'MentorLogout'])->name('Mentor.logout');
 });
+
 
 //staff
 Route::middleware(['auth', 'role:staff'])->group(function () {
@@ -117,20 +114,29 @@ Route::middleware(['auth', 'role:instansi'])->group(function () {
 
     Route::get('instansi/dashboard', [InstansiController::class, 'InstansiDashboard'])->name('instansi.dashboard');
     Route::get('instansi/logout', [InstansiController::class, 'InstansiLogout'])->name('instansi.logout');
-    // Route::get('admin/logout', [AdminController::class, 'AdminLogout'])->name('admin.logout');
+    Route::delete('instansi/pengajuan/delete/{id}',[PengajuanController::class,'destroyPengajuan'])->name('deletePengajuan');
+    Route::get('instansi/pengajuan/{id}',[PengajuanController::class,'detailPengajuan'])->name('detailPengajuanInst');
+    Route::post('instansi/pengajuan/{id}',[PengajuanController::class,'updatePengajuan'])->name('updatePengajuanInst');
 
+
+    Route::resource('instansi/pengajuan', PengajuanController::class);
 
 });
 
 Route::middleware(['auth', 'role:siswa'])->group(function () {
+
     Route::get('/dashboard', [UserController::class, 'UserDashboard'])->name('user.dashboard');
     Route::post('webcam', [UserController::class, 'store'])->name('webcam.capture');
     Route::get('/presensi', [UserController::class, 'UserPresensi'])->name('user.Presensi');
     Route::get('/presensis', [UserController::class, 'UserPresensis'])->name('user.Presensis');
 
     Route::get('/user/jurnal', [UserController::class, 'UserJurnal'])->name('user.jurnal');
+    Route::get('/user/jurnals', [UserController::class, 'CreateJurnal'])->name('create.jurnal');
+
+    Route::post('jurnal/store', [UserController::class, 'jurnalstore'])->name('user.store');
 
     Route::get('/user/materi', [UserController::class, 'UserMateri'])->name('user.materi');
     Route::get('user/logout', [UserController::class, 'UserLogout'])->name('user.logout');
     Route::post('/presensi/store/', [UserController :: class, 'store']);
 });
+
